CORS allow all
We are not relying on CORS as a defense mechanism. Using only
local storage on our front-end is both cleaner technically and
completely sidesteps the need for cors to stop browsers from doing
dangerous things.

Furthermore, we're totally ok with other websites relying on our API
right now. For the endpoints which are particularly security-sensitive,
we will use recaptcha explicitly which will perform domain
validation.