Add log routes, improve how integrations are passed around (#11)
This adds logging-related endpoints, and uses a lazy loading technique to avoid function signatures being blown up while still not initializing excess connections.

This meant it was relatively seamless to support read-only DB connections in the future, which could be a scaling improvement. In doing this this adjusted users.helper.get_auth_info_from_token_auth to no longer touch the last seen at or delete auth tokens which seem to have been leaked. The last seen at doesn't currently have a replacement:

Here is how we could replace the last_seen_at functionality so that we don't need those web-workers using write connections - store the value in the cache and send the auth id to a queue. If a bknd job worker sees the auth id in the queue and the value is still in the cache and the auth token is still valid, update the last seen at to the value in the cache). If the worker gets behind then we don't get last_seen_at values, which should be fine